+++ ddprof-lib/src/main/cpp/jvmSupport.cpp
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2026, Datadog, Inc.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include "jvmSupport.h"
+
+#include "asyncSampleMutex.h"
+#include "os.h"
+#include "thread.h"
+
+#include "hotspot/hotspotSupport.h"
+
+#include <jni.h>
+
+int JVMSupport::walkJavaStack(StackWalkRequest& request) {
+    if (VM::isHotspot()) {
+        return HotspotSupport::walkJavaStack(request);
+    } else if (VM::isOpenJ9() || VM::isZing()) {
+        return asyncGetCallTrace(request.frames, request.max_depth, request.ucontext);
+    }
+    assert(false && "Unsupported JVM");
+    return 0;
+}
+
+
+int JVMSupport::asyncGetCallTrace(ASGCT_CallFrame *frames, int max_depth, void* ucontext) {
+    JNIEnv *jni = VM::jni();
+    if (jni == nullptr) {
+        return 0;
+    }
+
+    AsyncSampleMutex mutex(ProfiledThread::currentSignalSafe());
+  
+    JitWriteProtection jit(false);
+    // AsyncGetCallTrace writes to ASGCT_CallFrame array
+    ASGCT_CallTrace trace = {jni, 0, frames};
+    VM::_asyncGetCallTrace(&trace, max_depth, ucontext);
+    return trace.num_frames;
+}
