+++ ddprof-lib/src/main/cpp/libraries.h
@@ -18,7 +18,7 @@ class Libraries {
   // This function will return the 'libjvm' on non-J9 VMs and the library with the given name on J9 VMs
   CodeCache *findJvmLibrary(const char *j9_lib_name);
   CodeCache *findLibraryByName(const char *lib_name);
-  CodeCache *findLibraryByAddress(const void *address);
+  CodeCache *findLibraryByAddress(const void *address) const;
 
   // Get library by index (used for remote symbolication unpacking)
   // Note: Parameter is uint32_t to match lib_index packing (17 bits = max 131K libraries)
