+++ ddprof-lib/src/main/cpp/profiler.cpp
@@ -15,8 +15,10 @@
 #include "flightRecorder.h"
 #include "itimer.h"
 #include "hotspot/vmStructs.h"
-#include "j9/j9Ext.h"
+#include "hotspot/hotspotSupport.h"
+#include "j9/j9Support.h"
 #include "j9/j9WallClock.h"
+#include "jvmSupport.h"
 #include "jvmThread.h"
 #include "libraryPatcher.h"
 #include "objectSampler.h"
@@ -28,7 +30,10 @@
 #include "symbols.h"
 #include "thread.h"
 #include "tsc.h"
+#include "utils.h"
 #include "wallClock.h"
+#include "frames.h"
+
 #include <algorithm>
 #include <dlfcn.h>
 #include <fstream>
@@ -57,54 +62,6 @@ static WallClockASGCT wall_asgct_engine;
 static J9WallClock j9_engine;
 static ITimer itimer;
 static CTimer ctimer;
-
-// Stack recovery techniques used to workaround AsyncGetCallTrace flaws.
-// Can be disabled with 'safemode' option.
-enum StackRecovery {
-  UNKNOWN_JAVA = (1 << 0),
-  POP_STUB = (1 << 1),
-  POP_METHOD = (1 << 2),
-  UNWIND_NATIVE = (1 << 3),
-  LAST_JAVA_PC = (1 << 4),
-  GC_TRACES = (1 << 5),
-  PROBE_SP = 0x100,
-};
-
-static inline int makeFrame(ASGCT_CallFrame *frames, jint type, jmethodID id) {
-  frames[0].bci = type;
-  frames[0].method_id = id;
-  return 1;
-}
-
-static inline int makeFrame(ASGCT_CallFrame *frames, jint type, uintptr_t id) {
-  return makeFrame(frames, type, (jmethodID)id);
-}
-
-static inline int makeFrame(ASGCT_CallFrame *frames, jint type,
-                            const char *id) {
-  return makeFrame(frames, type, (jmethodID)id);
-}
-
-void Profiler::addJavaMethod(const void *address, int length,
-                             jmethodID method) {
-  CodeHeap::updateBounds(address, (const char *)address + length);
-}
-
-void Profiler::addRuntimeStub(const void *address, int length,
-                              const char *name) {
-  _stubs_lock.lock();
-  _runtime_stubs.add(address, length, name, true);
-  _stubs_lock.unlock();
-
-  if (name[0] == 'I' && strcmp(name, "Interpreter") == 0) {
-    CodeHeap::setInterpreterStart(address);
-  } else if (strcmp(name, "call_stub") == 0) {
-    _call_stub_begin = address;
-    _call_stub_end = (const char *)address + length;
-  }
-
-  CodeHeap::updateBounds(address, (const char *)address + length);
-}
 
 void Profiler::onThreadStart(jvmtiEnv *jvmti, JNIEnv *jni, jthread thread) {
   ProfiledThread::initCurrentThread();
@@ -288,19 +245,6 @@ const char *Profiler::findNativeMethod(const void *address) {
     lib->binarySearch(address, &name);
   }
   return name;
-}
-
-CodeBlob *Profiler::findRuntimeStub(const void *address) {
-  return _runtime_stubs.findBlobByAddress(address);
-}
-
-bool Profiler::isAddressInCode(const void *pc, bool include_stubs) {
-  if (CodeHeap::contains(pc)) {
-    return CodeHeap::findNMethod(pc) != NULL &&
-           (include_stubs || !(pc >= _call_stub_begin && pc < _call_stub_end));
-  } else {
-    return _libs->findLibraryByAddress(pc) != NULL;
-  }
 }
 
 int Profiler::getNativeTrace(void *ucontext, ASGCT_CallFrame *frames,
@@ -491,273 +435,6 @@ int Profiler::convertNativeTrace(int native_frames, const void **callchain,
   return depth;
 }
 
-int Profiler::getJavaTraceAsync(void *ucontext, ASGCT_CallFrame *frames,
-                                int max_depth, StackContext *java_ctx,
-                                bool *truncated) {
-  // Workaround for JDK-8132510: it's not safe to call GetEnv() inside a signal
-  // handler since JDK 9, so we do it only for threads already registered in
-  // ThreadLocalStorage
-  VMThread *vm_thread = (VMThread*)JVMThread::current();
-  if (vm_thread == NULL || !vm_thread->isThreadAccessible()) {
-    Counters::increment(AGCT_NOT_REGISTERED_IN_TLS);
-    return 0;
-  }
-
-  JNIEnv *jni = VM::jni();
-  if (jni == NULL) {
-    // Not a Java thread
-    Counters::increment(AGCT_NOT_JAVA);
-    return 0;
-  }
-
-  StackFrame frame(ucontext);
-  uintptr_t saved_pc, saved_sp, saved_fp;
-  if (ucontext != NULL) {
-    saved_pc = frame.pc();
-    saved_sp = frame.sp();
-    saved_fp = frame.fp();
-
-    if (saved_pc >= (uintptr_t)_call_stub_begin &&
-        saved_pc < (uintptr_t)_call_stub_end) {
-      // call_stub is unsafe to walk
-      frames->bci = BCI_ERROR;
-      frames->method_id = (jmethodID) "call_stub";
-      return 1;
-    }
-
-    if (!VMStructs::isSafeToWalk(saved_pc)) {
-      frames->bci = BCI_NATIVE_FRAME;
-      CodeBlob *codeBlob =
-          VMStructs::libjvm()->findBlobByAddress((const void *)saved_pc);
-      if (codeBlob) {
-        frames->method_id = (jmethodID)codeBlob->_name;
-      } else {
-        frames->method_id = (jmethodID) "unknown_unwalkable";
-      }
-      return 1;
-    }
-  } else {
-    return 0;
-  }
-
-  int state = vm_thread->state();
-  // from OpenJDK
-  // https://github.com/openjdk/jdk/blob/7455bb23c1d18224e48e91aae4f11fe114d04fab/src/hotspot/share/utilities/globalDefinitions.hpp#L1030
-  /*
-  enum JavaThreadState {
-      _thread_uninitialized     =  0, // should never happen (missing
-  initialization) _thread_new               =  2, // just starting up, i.e., in
-  process of being initialized _thread_new_trans         =  3, // corresponding
-  transition state (not used, included for completeness) _thread_in_native =  4,
-  // running in native code _thread_in_native_trans   =  5, // corresponding
-  transition state _thread_in_vm             =  6, // running in VM
-      _thread_in_vm_trans       =  7, // corresponding transition state
-      _thread_in_Java           =  8, // running in Java or in stub code
-      _thread_in_Java_trans     =  9, // corresponding transition state (not
-  used, included for completeness) _thread_blocked           = 10, // blocked in
-  vm _thread_blocked_trans     = 11, // corresponding transition state
-      _thread_max_state         = 12  // maximum thread state+1 - used for
-  statistics allocation
-  };
-   */
-  bool in_java = (state == 8 || state == 9);
-  if (in_java && java_ctx->sp != 0) {
-    // skip ahead to the Java frames before calling AGCT
-    frame.restore((uintptr_t)java_ctx->pc, java_ctx->sp, java_ctx->fp);
-  } else if (state != 0) {
-    VMJavaFrameAnchor* a = vm_thread->anchor();
-    if (a == nullptr || a->lastJavaSP() == 0) {
-      // we haven't found the top Java frame ourselves, and the lastJavaSP wasn't
-      // recorded either when not in the Java state, lastJava ucontext will be
-      // used by AGCT
-      Counters::increment(AGCT_NATIVE_NO_JAVA_CONTEXT);
-      return 0;
-    }
-  }
-  bool blocked_in_vm = (state == 10 || state == 11);
-  // avoid unwinding during deoptimization
-  if (blocked_in_vm && vm_thread->osThreadState() == OSThreadState::RUNNABLE) {
-    Counters::increment(AGCT_BLOCKED_IN_VM);
-    return 0;
-  }
-
-  JitWriteProtection jit(false);
-  // AsyncGetCallTrace writes to ASGCT_CallFrame array
-  ASGCT_CallTrace trace = {jni, 0, frames};
-  VM::_asyncGetCallTrace(&trace, max_depth, ucontext);
-
-  if (trace.num_frames > 0) {
-    frame.restore(saved_pc, saved_sp, saved_fp);
-    return trace.num_frames;
-  }
-
-  if ((trace.num_frames == ticks_unknown_Java ||
-       trace.num_frames == ticks_not_walkable_Java) &&
-      !(_safe_mode & UNKNOWN_JAVA) && ucontext != NULL) {
-    CodeBlob *stub = NULL;
-    _stubs_lock.lockShared();
-    if (_runtime_stubs.contains((const void *)frame.pc())) {
-      stub = findRuntimeStub((const void *)frame.pc());
-    }
-    _stubs_lock.unlockShared();
-
-    if (stub != NULL) {
-      if (_cstack != CSTACK_NO) {
-        max_depth -= makeFrame(trace.frames++, BCI_NATIVE_FRAME, stub->_name);
-      }
-      if (!(_safe_mode & POP_STUB) &&
-          frame.unwindStub((instruction_t *)stub->_start, stub->_name) &&
-          isAddressInCode((const void *)frame.pc())) {
-        VM::_asyncGetCallTrace(&trace, max_depth, ucontext);
-      }
-    } else if (VMStructs::hasMethodStructs()) {
-      VMNMethod *nmethod = CodeHeap::findNMethod((const void *)frame.pc());
-      if (nmethod != NULL && nmethod->isNMethod() && nmethod->isAlive()) {
-        VMMethod *method = nmethod->method();
-        if (method != NULL) {
-          jmethodID method_id = method->id();
-          if (method_id != NULL) {
-            max_depth -= makeFrame(trace.frames++, 0, method_id);
-          }
-          if (!(_safe_mode & POP_METHOD) && frame.unwindCompiled(nmethod) &&
-              isAddressInCode((const void *)frame.pc())) {
-            VM::_asyncGetCallTrace(&trace, max_depth, ucontext);
-          }
-          if ((_safe_mode & PROBE_SP) && trace.num_frames < 0) {
-            if (method_id != NULL) {
-              trace.frames--;
-            }
-            for (int i = 0; trace.num_frames < 0 && i < PROBE_SP_LIMIT; i++) {
-              frame.sp() += sizeof(void*);
-              VM::_asyncGetCallTrace(&trace, max_depth, ucontext);
-            }
-          }
-        }
-      } else if (nmethod != NULL) {
-        if (_cstack != CSTACK_NO) {
-          max_depth -=
-              makeFrame(trace.frames++, BCI_NATIVE_FRAME, nmethod->name());
-        }
-        if (!(_safe_mode & POP_STUB) &&
-            frame.unwindStub(NULL, nmethod->name()) &&
-            isAddressInCode((const void *)frame.pc())) {
-          VM::_asyncGetCallTrace(&trace, max_depth, ucontext);
-        }
-      }
-    }
-  } else if (trace.num_frames == ticks_unknown_not_Java &&
-             !(_safe_mode & LAST_JAVA_PC)) {
-    VMJavaFrameAnchor* anchor = vm_thread->anchor();
-    if (anchor == NULL) return 0;
-    uintptr_t sp = anchor->lastJavaSP();
-    const void* pc = anchor->lastJavaPC();
-    if (sp != 0 && pc == NULL) {
-      // We have the last Java frame anchor, but it is not marked as walkable.
-      // Make it walkable here
-      pc = ((const void**)sp)[-1];
-      anchor->setLastJavaPC(pc);
-
-      VMNMethod *m = CodeHeap::findNMethod(pc);
-      if (m != NULL) {
-        // AGCT fails if the last Java frame is a Runtime Stub with an invalid
-        // _frame_complete_offset. In this case we patch _frame_complete_offset
-        // manually
-        if (!m->isNMethod() && m->frameSize() > 0 &&
-            m->frameCompleteOffset() == -1) {
-          m->setFrameCompleteOffset(0);
-        }
-        VM::_asyncGetCallTrace(&trace, max_depth, ucontext);
-      } else if (_libs->findLibraryByAddress(pc) != NULL) {
-        VM::_asyncGetCallTrace(&trace, max_depth, ucontext);
-      }
-
-      anchor->setLastJavaPC(nullptr);
-    }
-  } else if (trace.num_frames == ticks_not_walkable_not_Java &&
-             !(_safe_mode & LAST_JAVA_PC)) {
-    VMJavaFrameAnchor* anchor = vm_thread->anchor();
-    if (anchor == NULL) return 0;
-    uintptr_t sp = anchor->lastJavaSP();
-    const void* pc = anchor->lastJavaPC();
-    if (sp != 0 && pc != NULL) {
-      // Similar to the above: last Java frame is set,
-      // but points to a Runtime Stub with an invalid _frame_complete_offset
-      VMNMethod *m = CodeHeap::findNMethod(pc);
-      if (m != NULL && !m->isNMethod() && m->frameSize() > 0 &&
-          m->frameCompleteOffset() == -1) {
-        m->setFrameCompleteOffset(0);
-        VM::_asyncGetCallTrace(&trace, max_depth, ucontext);
-      }
-    }
-  } else if (trace.num_frames == ticks_GC_active && !(_safe_mode & GC_TRACES)) {
-    VMJavaFrameAnchor* anchor = vm_thread->anchor();
-    if (anchor == NULL || anchor->lastJavaSP() == 0) {
-      // Do not add 'GC_active' for threads with no Java frames, e.g. Compiler
-      // threads
-      frame.restore(saved_pc, saved_sp, saved_fp);
-      return 0;
-    }
-  }
-
-  frame.restore(saved_pc, saved_sp, saved_fp);
-
-  if (trace.num_frames > 0) {
-    return trace.num_frames + (trace.frames - frames);
-  }
-
-  const char *err_string = asgctError(trace.num_frames);
-  if (err_string == NULL) {
-    // No Java stack, because thread is not in Java context
-    return 0;
-  }
-
-  atomicIncRelaxed(_failures[-trace.num_frames]);
-  trace.frames->bci = BCI_ERROR;
-  trace.frames->method_id = (jmethodID)err_string;
-  return trace.frames - frames + 1;
-}
-
-void Profiler::fillFrameTypes(ASGCT_CallFrame *frames, int num_frames,
-                              VMNMethod *nmethod) {
-  if (nmethod->isNMethod() && nmethod->isAlive()) {
-    VMMethod *method = nmethod->method();
-    if (method == NULL) {
-      return;
-    }
-
-    jmethodID current_method_id = method->id();
-    if (current_method_id == NULL) {
-      return;
-    }
-
-    // Mark current_method as COMPILED and frames above current_method as
-    // INLINED
-    for (int i = 0; i < num_frames; i++) {
-      if (frames[i].method_id == NULL || frames[i].bci <= BCI_NATIVE_FRAME) {
-        break;
-      }
-      if (frames[i].method_id == current_method_id) {
-        int level = nmethod->level();
-        frames[i].bci = FrameType::encode(
-            level >= 1 && level <= 3 ? FRAME_C1_COMPILED : FRAME_JIT_COMPILED,
-            frames[i].bci);
-        for (int j = 0; j < i; j++) {
-          frames[j].bci = FrameType::encode(FRAME_INLINED, frames[j].bci);
-        }
-        break;
-      }
-    }
-  } else if (nmethod->isInterpreter()) {
-    // Mark the first Java frame as INTERPRETED
-    for (int i = 0; i < num_frames; i++) {
-      if (frames[i].bci > BCI_NATIVE_FRAME) {
-        frames[i].bci = FrameType::encode(FRAME_INTERPRETED, frames[i].bci);
-        break;
-      }
-    }
-  }
-}
-
 u64 Profiler::recordJVMTISample(u64 counter, int tid, jthread thread, jint event_type, Event *event, bool deferred) {
   // Protect JVMTI sampling operations to prevent signal handler interference
   CriticalSection cs;
@@ -859,34 +536,13 @@ void Profiler::recordSample(void *ucontext, u64 counter, int tid,
     ASGCT_CallFrame *native_stop = frames + num_frames;
     num_frames += getNativeTrace(ucontext, native_stop, event_type, tid,
                                  &java_ctx, &truncated, lock_index);
-    if (num_frames < _max_stack_depth) {
-      int max_remaining = _max_stack_depth - num_frames;
-      if (_features.mixed) {
-        int vm_start = num_frames;
-        int vm_frames = StackWalker::walkVM(ucontext, frames + vm_start, max_remaining, _features, eventTypeFromBCI(event_type), lock_index, &truncated);
-        num_frames += vm_frames;
-      } else if (event_type == BCI_CPU || event_type == BCI_WALL) {
-        if (_cstack >= CSTACK_VM) {
-          int vm_start = num_frames;
-          int vm_frames = StackWalker::walkVM(ucontext, frames + vm_start, max_remaining, _features, eventTypeFromBCI(event_type), lock_index, &truncated);
-          num_frames += vm_frames;
-        } else {
-          // Async events
-          AsyncSampleMutex mutex(ProfiledThread::currentSignalSafe());
-          int java_frames = 0;
-          if (mutex.acquired()) {
-            java_frames = getJavaTraceAsync(ucontext, frames + num_frames, max_remaining, &java_ctx, &truncated);
-            if (java_frames > 0 && java_ctx.pc != NULL && VMStructs::hasMethodStructs()) {
-              VMNMethod* nmethod = CodeHeap::findNMethod(java_ctx.pc);
-              if (nmethod != NULL) {
-                fillFrameTypes(frames + num_frames, java_frames, nmethod);
-              }
-            }
-          }
-          num_frames += java_frames;
-        }
-      }
+    int max_remaining = _max_stack_depth - num_frames;
+    if (max_remaining > 0) {
+      // Walk Java frames if we have room, but only for mixed mode or CPU/Wall events with cstack enabled. For async events, we want to avoid walking Java frames in the signal handler if possible, since it can lead to deadlocks. Instead, we'll try to get the Java trace asynchronously after the signal handler returns.
+      StackWalkRequest request = {_features, event_type, _cstack, lock_index, ucontext, frames + num_frames, max_remaining, &java_ctx, &truncated};
+      num_frames += JVMSupport::walkJavaStack(request);
     }
+  
     if (num_frames == 0) {
       num_frames += makeFrame(frames + num_frames, BCI_ERROR, "no_Java_frame");
     }
@@ -1025,6 +681,20 @@ void *Profiler::dlopen_hook(const char *filename, int flags) {
   return result;
 }
 
+const char* Profiler::cstack() const {
+  switch (_cstack) {
+    case CSTACK_DEFAULT: return "default";
+    case CSTACK_NO: return "no";
+    case CSTACK_FP: return "fp";
+    case CSTACK_DWARF: return "dwarf";
+    case CSTACK_LBR: return "lbr";
+    case CSTACK_VM: {
+      return _features.mixed ? "vmx" : "vm";
+    }
+    default: return "default";
+  }
+}
+
 void Profiler::switchLibraryTrap(bool enable) {
   if (_dlopen_entry == NULL) {
     return;  // Not initialized yet, nothing to do
@@ -1083,7 +753,7 @@ int Profiler::crashHandlerInternal(int signo, siginfo_t *siginfo, void *ucontext
   // Reentrancy protection: use TLS-based tracking if available.
   // If TLS is not available, we can only safely handle faults that we can
   // prove are from our protected code paths (checked via sameStack heuristic
-  // in StackWalker::checkFault). For anything else, we must chain immediately
+  // in HotspotSupport::checkFault). For anything else, we must chain immediately
   // to avoid claiming faults that aren't ours.
   bool have_tls_protection = false;
   if (thrd != nullptr) {
@@ -1094,7 +764,7 @@ int Profiler::crashHandlerInternal(int signo, siginfo_t *siginfo, void *ucontext
     have_tls_protection = true;
   }
   // If thrd == nullptr, we proceed but with limited handling capability.
-  // Only StackWalker::checkFault (which has its own sameStack fallback)
+  // Only HotspotSupport::checkFault (which has its own sameStack fallback)
   // and the JDK-8313796 workaround can safely handle faults without TLS.
 
   StackFrame frame(ucontext);
@@ -1119,11 +789,11 @@ int Profiler::crashHandlerInternal(int signo, siginfo_t *siginfo, void *ucontext
   if (VM::isHotspot()) {
     // the following checks require vmstructs and therefore HotSpot
 
-    // StackWalker::checkFault has its own fallback for when TLS is unavailable:
+    // HotspotSupport::checkFault has its own fallback for when TLS is unavailable:
     // it uses sameStack() heuristic to check if we're in a protected stack walk.
     // If the fault is from our protected walk, it will longjmp and never return.
     // If it returns, the fault wasn't from our code.
-    StackWalker::checkFault(thrd);
+    HotspotSupport::checkFault(thrd);
 
     // Workaround for JDK-8313796 if needed. Setting cstack=dwarf also helps
     if (_need_JDK_8313796_workaround &&
@@ -1229,8 +899,8 @@ Engine *Profiler::selectCpuEngine(Arguments &args) {
     return &noop_engine;
   } else if (args._cpu >= 0 || strcmp(args._event, EVENT_CPU) == 0) {
     if (VM::isOpenJ9()) {
-      if (!J9Ext::shouldUseAsgct() || !J9Ext::can_use_ASGCT()) {
-        if (!J9Ext::is_jvmti_jmethodid_safe()) {
+      if (!J9Support::shouldUseAsgct() || !J9Support::can_use_ASGCT()) {
+        if (!J9Support::is_jvmti_jmethodid_safe()) {
           LOG_WARN("Safe jmethodID access is not available on this JVM. Using "
                     "CPU profiler on your own risk. Use -XX:+KeepJNIIDs=true JVM "
                     "flag to make access to jmethodIDs safe, if your JVM supports it");
@@ -1261,8 +931,8 @@ Engine *Profiler::selectWallEngine(Arguments &args) {
     return &noop_engine;
   }
   if (VM::isOpenJ9()) {
-    if (args._wallclock_sampler == JVMTI || !J9Ext::shouldUseAsgct() || !J9Ext::can_use_ASGCT()) {
-      if (!J9Ext::is_jvmti_jmethodid_safe()) {
+    if (args._wallclock_sampler == JVMTI || !J9Support::shouldUseAsgct() || !J9Support::can_use_ASGCT()) {
+      if (!J9Support::is_jvmti_jmethodid_safe()) {
         LOG_WARN("Safe jmethodID access is not available on this JVM. Using "
                   "wallclock profiler on your own risk. Use -XX:+KeepJNIIDs=true JVM "
                   "flag to make access to jmethodIDs safe, if your JVM supports it");
